requirements.txt
Pillow==7.0.0
python-resize-image==1.1.19
instabot==0.112.0